+++ packages/vtable/src/core/BaseTable.ts
@@ -1105,6 +1105,12 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
       this._clearRowRangeHeightsMap(row);
     }
   }
+
+  setRowHeight(row: number, height: number) {
+    this.scenegraph.setRowHeight(row, height);
+    this.internalProps._heightResizedRowMap.add(row); // add resize tag
+  }
+
   /**
    * 获取指定行范围的总高度
    * @param startCol
@@ -1293,6 +1299,11 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     if (!skipCheckFrozen) {
       this.stateManager.checkFrozen();
     }
+  }
+
+  setColWidth(col: number, width: number) {
+    this.scenegraph.setColWidth(col, width);
+    this.internalProps._widthResizedColMap.add(col); // add resize tag
   }
 
   /**
@@ -2670,11 +2681,19 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     const { scrollLeft, scrollTop } = this;
     cellRanges.forEach((cellRange: CellRange, index: number) => {
       if (cellRange.start.col === cellRange.end.col && cellRange.start.row === cellRange.end.row) {
-        this.stateManager.updateSelectPos(cellRange.start.col, cellRange.start.row, false, index >= 1);
+        this.stateManager.updateSelectPos(
+          cellRange.start.col,
+          cellRange.start.row,
+          false,
+          index >= 1,
+          false,
+          false,
+          true
+        );
       } else {
         this.stateManager.updateSelectPos(cellRange.start.col, cellRange.start.row, false, index >= 1);
         this.stateManager.updateInteractionState(InteractionState.grabing);
-        this.stateManager.updateSelectPos(cellRange.end.col, cellRange.end.row, false, index >= 1);
+        this.stateManager.updateSelectPos(cellRange.end.col, cellRange.end.row, false, index >= 1, false, false, true);
       }
       this.stateManager.endSelectCells(false);
       this.stateManager.updateInteractionState(InteractionState.default);